requirements/local.txt
# Add test and quality tools over here
-r base.txt

verage==4.5.2
django-coverage-plugin==1.6.0

django-silk==3.0.1co